example/requirements.txt
wheel
asyncio
asyncio-nats-client
